+++ CHANGELOG.md
@@ -3,16 +3,27 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-# [1.4.0](https://github.com/nativescript-community/ui-webview/compare/v1.3.5...v1.4.0) (2022-08-18)
+## [1.4.2](https://github.com/nativescript-community/ui-webview/compare/v1.4.1...v1.4.2) (2022-09-02)
 
+### Bug Fixes
+
+* rename trace category ([33d1f60](https://github.com/nativescript-community/ui-webview/commit/33d1f6058a4defd29eedbf390465504faab3e1d9))
+
+## [1.4.1](https://github.com/nativescript-community/ui-webview/compare/v1.4.0...v1.4.1) (2022-09-02)
+
+### Bug Fixes
+
+* **android:** cacheMode improvements ([01aeb59](https://github.com/nativescript-community/ui-webview/commit/01aeb595192f5d3e429c5b4e1e098739628c93ae))
+
+### Features
+
+* **webview:** `useWideViewPort` for android ([88f9e38](https://github.com/nativescript-community/ui-webview/commit/88f9e389e7f01e24198e045694edef12bed5799e))
+
+# [1.4.0](https://github.com/nativescript-community/ui-webview/compare/v1.3.5...v1.4.0) (2022-08-18)
 
 ### Features
 
 * **webview:** iOS support for `limitsNavigationsToAppBoundDomains` ([ab1435b](https://github.com/nativescript-community/ui-webview/commit/ab1435b4b1f2e76836df8534f5cf5ef01240bdcf))
-
-
-
-
 
 ## [1.3.5](https://github.com/nativescript-community/ui-webview/compare/v1.3.4...v1.3.5) (2022-05-25)